requirements.txt
numpy>=1.21.0
nibabel>=3.2.0
scipy>=1.7.0
pyyaml>=6.0 
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tensorboard>=2.14.0
tqdm>=4.65.0 